Move site_packages to fawtydeps utils.
Reorder package search conditions and simplify one for Windows to resemble one for POSIX more.
